[IMP] web: convert core/field_service tests to hoot
task-id: 3705027
